#511 [MODIFY] Submit analytics at onViewCreated
